# fix(utils): speedup random string generation