Results:
```
pkg: github.com/scylladb/gemini/pkg/utils
cpu: 12th Gen Intel(R) Core(TM) i9-12900HK
BenchmarkUtilsRandStringNew100
BenchmarkUtilsRandStringNew100-20     	 9158696	       157.4 ns/op
BenchmarkUtilsRandStringNew1000
BenchmarkUtilsRandStringNew1000-20    	 2407350	       592.3 ns/op
BenchmarkUtilsRandStringOld100
BenchmarkUtilsRandStringOld100-20      	 1878685	       666.5 ns/op
BenchmarkUtilsRandStringOld1000
BenchmarkUtilsRandStringOld1000-20     	  174566	      6064 ns/op
PASS
```